Re-org and rename folders and make server.js more concise
