+++ package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node ./server.js",
-    "start:dev": "nodemon ./server.js",
+    "start": "node ./src/server.js",
+    "start:dev": "nodemon ./src/server.js",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix"
   },
@@ -55,6 +55,7 @@
     "eslint-config-airbnb": "^16.1.0",
     "eslint-plugin-import": "^2.11.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.7.0"
+    "eslint-plugin-react": "^7.7.0",
+    "nodemon": "^1.17.4"
   }
 }
